Refactor large services module into smaller separate modules
