+++ src/js/init.js
@@ -406,10 +406,9 @@ function init() {
     }
   });
 
-  promise.then(function(taxids) {
+  promise.then(function() {
+    taxids = ideo.config.taxids;
     taxid = taxids[0];
-    ideo.config.taxid = taxid;
-    ideo.config.taxids = taxids;
 
     var assemblies,
       bandFileName;
